architecture/implementations/chat_gateway/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
httpx==0.25.1
redis==5.0.1
pyjwt==2.8.0
python-multipart==0.0.6
prometheus-client==0.18.0
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-exporter-otlp==1.20.0 